% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessAquarius.R
\name{AccessAquarius}
\alias{AccessAquarius}
\title{Access Aquarius}
\usage{
AccessAquarius(hostname, username, password)
}
\arguments{
\item{hostname}{Defaults to "https://aquarius.nps.gov/aquarius" if the server changes this may leave the option for continued use.}

\item{username}{Defaults to "aqreadonly", which should work for downloading.  This parameter is retained in the event you need to use your specific log in.}

\item{password}{Defaults to "aqreadonly", which should work for downloading.  This parameter is retained in the event you need to use your specific log in.}
}
\description{
Establishes connection to Aquarius server for future functions to search and download data. This function require NPS network access (or VPN).
}
\examples{
AccessAquarius()

}
