% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchDatasetIDs.R
\name{searchDatasetIDs}
\alias{searchDatasetIDs}
\title{Search Dataset IDs}
\usage{
searchDatasetIDs(site, searchterm, removeterms)
}
\arguments{
\item{site}{Name of the site to search for datasets.}

\item{searchterm}{(optional) text that will be used to find dataset IDs that contain this value if no search term is provided all dataset IDs will be returned.  Note that this term is case sensitive.}

\item{removeterms}{(optional) text that will be used to limit the returned dataset IDs. Any IDs that contain this term will not be returned.  Note that this term is case sensitive.}
}
\value{
An object of type character with a value for each of the dataset IDs at that site.
}
\description{
This function takes searches a specific site for dataset IDs that contain the value provided. It is useful for identifying the correct search terms to find the right datasets names.
}
\examples{
Pulling up all the dataset IDs at the site "DENA-023"
searchDatasetIDs("DENA-023")

Pulling up the dataset IDs that contain "Depth@" at the site "YUCH-005"
searchDatasetIDs("YUCH-005", "Depth@")

Pulling up the dataset IDs that contain "pH@" but does not contain "synop" at site "BELA-088"
searchDatasetIDs("BELA-088",searchterm = "pH@", removeterms = "synop")

Pulling up the dataset IDs at the site "KOVA-085" that do not contain "ynop"
searchDatasetIDs("KOVA-085",  removeterms = "ynop")

}
