% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableJoin.R
\name{tablejoin}
\alias{tablejoin}
\title{Table Join}
\usage{
tablejoin(table1, table2)
}
\arguments{
\item{table1}{First Table}

\item{table2}{Second Table}
}
\description{
This function joins tables using rbind but it will add blank columns to make the columns match if there are discrepancies. Soon to be replaced by dplyr::row_bind.
}
\examples{
tablea <- data.frame(Letter=c("A", "B", "C"), Integer=c(1,2,3))
tableb <- data.frame(Letter=c("A", "D", "E"), Animal=c("Cat", "Dog", "Snake"))
tablejoin(tablea, tableb)
  Letter Integer Animal
1      A       1
2      B       2
3      C       3
4      A            Cat
5      D            Dog
6      E          Snake

}
