% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadaquariusdata.R
\name{downloadaquariusdata}
\alias{downloadaquariusdata}
\title{Download Aquarius Data}
\usage{
downloadaquariusdata(datasets, sites, startyear, endyear, site.rename)
}
\arguments{
\item{datasets}{is a data frame in the format of the output from the finddatasets() function.}

\item{sites}{(optional) is a list of sites to download data for. This list cannot include any sites not in datasets table but it does not need to include all of those sites. If it is ignored all sites from the parameter table will be included.}

\item{startyear}{(optional) is a single value of the year you wish to start your time series. Data prior to January 1st of this year will not be downloaded.}

\item{endyear}{(optional) is a single value of the year you wish to end your time series. Data after December 31st of this year will not be downloaded.}

\item{site.rename}{(optional) is a vector of names to replace the site names in the output.  This argument allows the site ID in the output to not exactly match the site ID in aquarius.}
}
\value{
A data frame with all the data from the sites and parameters listed in the arguments
}
\description{
This function downloads the data from all datasets listed in the parameter table.
}
\examples{
#ParameterTable <- finddatasets(sites=c("BELA-088", "BELA-108", "KOVA-085", "YUCH-005"),
#                               parameters = c("LakeDepth", "SedimentTemperatureC","DissolvedOxygen(ODOconc)","pH","SpecificConductance(SpCond)","Temperature"),
#                               searchterms = c("Pressure_LakeDepth@", "Pressure_BottomTemp@", "DOconc@", "Sonde_pH@", "Sonde_SpCond@", "Sonde_Temp@"),
#                                removeterms = "synoptic")

# Downloading the lake data fully based on the defaults
DownloadedData <- downloadaquariusdata(ParameterTable)

# Downloading the lake data only from the site "BELA-088"
DownloadedData <- downloadaquariusdata(ParameterTable, sites = "BELA-088")

# Downloading the lake data only from the site "BELA-108" but renaming it "Site1"
DownloadedData <- downloadaquariusdata(ParameterTable, sites = "BELA-108", site.rename = "Site1")

# Downloading the lake data but only from the sites "KOVA-085" and "YUCH-005" but renaming the KOVA site as "Site1" and the YUCH site as "Site2" and "Site3".
DownloadedData <- downloadaquariusdata(ParameterTable, sites = c("KOVA-085","YUCH-005", "YUCH-005") , site.rename = c("Site1","Site2", "Site3"))

}
